Make UnixDomainSocketLibrary graalvm friendly
The graalvm native images have problems with jna registration in a
static initializer. To work around this, I moved the implementation of
the native code into a separate class. For some reason, using the
ByteBuffer jna implementation of read and write lead to bad address
exceptions. To workaround this without breaking binary compatibility, we
can just pass in the backing array of the byte buffer to the jna
implementation that takes a byte array rather than a byte buffer.