Make windows named pipe security level configurable
In 986a33cd, we started adding access control to the named pipes created
for windows server sockets. This was designed to protect snooping across
different user accounts. The security choice made was to restrict client
connections to the login session. The drawback to this choice was that
it prevents remote ssh connections to the same account from accessing
the named pipe. This commit makes the security level configurable and
keeps the default security level to be the logon dacl. sbt, which is the
primary downstream consumer of this library, will likely be updated to
use the owner dacl by default while allowing users to also choose to use
the more restrictive logon dacl option or forgo security entirely if they wish.